Refactor ManagedWebSocket out of ClientWebSocket
To make ManagedWebSocket shareable with ASP.NET.
