+++ src/System.Net.WebSockets.Client/src/System.Net.WebSockets.Client.csproj
@@ -112,7 +112,9 @@
     <Compile Include="System\Net\WebSockets\WinRTWebSocket.cs" />
   </ItemGroup>
   <ItemGroup Condition=" '$(TargetsUnix)' == 'true' ">
-    <Compile Include="System\Net\WebSockets\WebSocketHandle.Unix.cs" />
+    <Compile Include="System\Net\Sockets\AsyncEventArgsNetworkStream.cs" />
+    <Compile Include="System\Net\WebSockets\WebSocketHandle.Managed.cs" />
+    <Compile Include="System\Net\WebSockets\ManagedWebSocket.cs" />
     <Compile Include="$(CommonPath)\System\StringExtensions.cs">
       <Link>Common\System\StringExtensions.cs</Link>
     </Compile>
